requirements.txt
fastapi
uvicorn
jinja2
python-multipart
web3
qrcode[pil]
httpx
python-dotenv
requests